requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
PyGithub==1.59.1
atlassian-python-api==3.41.0
slack-sdk==3.26.1
python-multipart==0.0.6
aiofiles==23.2.1